Fix Velocity plugin
AntiVPN tries to load dependencies in an independent class loader. This is not well-supported in Velocity, but Velocity allows any plugin to inject files into their own classpath. As a quick fix, implement a fake URLClassLoader that just forwards to Velocity's PluginManager#addToClasspath(). A longer-term fix would be to use a solution similar to LuckPerms's solution.

This fix has only been tested on Linux, I'm not sure how well Windows will respond.